invitations/requirements.txt
Django==2.2
django-qr-code==1.0.0
django-smtp-ssl==1.0
djangorestframework==3.10.3
django-recaptcha==2.0.5
psycopg2==2.8.4
python-dotenv==0.10.3
pytz==2019.3
qrcode==6.1
six==1.13.0
sqlparse==0.3.0